requirements.txt
langchain==0.1.14
langchain_community==0.0.31
python-dotenv==1.0.1
